update ProjectId on collision
