+++ server/portal/apps/projects/managers/base.py
@@ -13,6 +13,7 @@ from portal.libs.agave.models.systems.storage import StorageSystem
 from portal.libs.elasticsearch.docs.base import IndexedProject
 from portal.apps.projects.models import Project, ProjectId, ProjectSystemSerializer
 from portal.apps.projects.serializers import MetadataJSONSerializer
+from portal.apps.projects.models.utils import get_latest_project_storage, get_latest_project_directory
 
 
 # pylint: disable=invalid-name
@@ -154,12 +155,30 @@ class ProjectsManager(object):
             prefix=settings.PORTAL_PROJECTS_ID_PREFIX,
             prj_id=ProjectId.next_id()
         )
-        prj = Project.create(
-            self.user.agave_oauth.client,
-            title,
-            project_id,
-            self.user
-        )
+        try:
+            prj = Project.create(
+                self.user.agave_oauth.client,
+                title,
+                project_id,
+                self.user
+            )
+        except ValueError:
+            logger.info('Project with id: {} already exists'.format(project_id))
+            latest_storage_system_id = get_latest_project_storage()
+            latest_project_id = get_latest_project_directory()
+            max_value_found = max(latest_storage_system_id, latest_project_id, 0)
+            logger.info('Updating to ProjectId latest storage system id: {}'.format(max_value_found))
+            ProjectId.update(max_value_found)
+            project_id = '{prefix}-{prj_id}'.format(
+                prefix=settings.PORTAL_PROJECTS_ID_PREFIX,
+                prj_id=ProjectId.next_id()
+            )
+            prj = Project.create(
+                self.user.agave_oauth.client,
+                title,
+                project_id,
+                self.user
+            )
         prj.storage.update_role(
             self.user.username,
             'ADMIN'
